4.21 - step9

+++ osa4/controllers/blogs.js
@@ -57,8 +57,22 @@ blogsRouter.post('/', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+ const blog =  await Blog.findById(request.params.id)
+ const userid = user.id
+ const user = await User.findById(decodedToken.id)
+
+  if ( blog.user.toString() === userid.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'ei oikeuksia poistaa' })
+  }
 })
 
 module.exports = blogsRouter
